requirements.txt
python-telegram-bot==20.7
gspread==5.12.0
google-auth==2.23.4
matplotlib==3.8.2
seaborn==0.13.0
pandas==2.1.4
python-dotenv==1.0.0
pytz==2023.3 